refactor(tv): extract useSessionPlayback hook + shared VariantRow/panel styles
Review-followup structural refactors (behavior-preserving):
- VideoPlayer.tsx 1591->1410 lines via useSessionPlayback hook
- shared VariantRow + panelStyles dedupes language/in-player rows
- activeVariantMediaState exposed on WatchSession; panels stop reconstructing
